replace Strapi types with Sanity types
Remove Strapi-era type files (album.ts, carousel.ts, strapi.ts) and add
types/sanity.ts with SanityAlbum, SanityPhoto, SanityImageField,
SanityImageAsset, and CarouselEntry. Update types/image.ts accordingly.
