--- types/carousel.ts
@@ -1,3 +1,0 @@
-export interface CarouselEntry {
-  readonly imageId: number; // Strapi media library image id
-}
