+++ types/image.ts
@@ -1,8 +1,5 @@
-import type { StrapiImageFormats } from "./strapi.ts";
-
 export interface ImageProps {
   readonly alt: string;
-  readonly formats?: StrapiImageFormats | null;
   readonly height: number;
   readonly src: string;
   readonly width: number;
